Conflicts:
	cal.pot
	en_GB.po
	en_GB.utf8.po
	en_US.po
	es_ES.iso88591.po
	es_ES.utf8.po
	fi_FI.po
	fi_FI.utf8.po
	leap.c
	leap.h
	nextday.c
	weekday.c

+++ cal.c
@@ -20,6 +20,10 @@
 #include "nextday.h"
 #include "weekday.h"
 
+#include "nextday.h"
+#include "weekday.h"
+#include "leap.h"
+
 #define NMONTHS     12
 #define NDAYS       7
 
@@ -67,7 +71,7 @@ size_t mbstrlen(char *s)
     mbstate_t st = {0};
     int i = 0;
 
-    size_t n;
+    ssize_t n;
 #if 0
     printf(D("[%s]:"), s);
 #endif
@@ -216,12 +220,12 @@ void long_format(int year)
     int day[3], offset[3], month[3], i, j, k, l;
     int hl[3], ml[3], tl[3];
 	char str [10];
-    size_t n;
+    ssize_t n;
 
 	n = sprintf(str, "%d", year);
 
-    printf("\n\n%*s\n", 38 + n/2, str);
-    printf("%*.*s\n", 38 + n/2, n, "============");
+    printf("\n\n%*s\n", 38 + (int)n/2, str);
+    printf("%*.*s\n", 38 + (int)n/2, (int)n, "============");
 
     printf("\n--------------------------------------"
         	 "--------------------------------------\n");
@@ -260,7 +264,7 @@ void long_format(int year)
 			for (k = 0; k < 7; k++) {
 				printf("%s%.*s",
 					k ? " " : "",
-                    days[(k + fdow)%7].di_l2,
+                    (int)days[(k + fdow)%7].di_l2,
                     days[(k + fdow)%7].di_name);
 			} /* for */
 		} /* for */
@@ -316,7 +320,7 @@ void short_format(int month, int year)
 	for (i = 0; i < 7; i++)
 		printf("%s%.*s",
                 i ? " " : "",
-                days[(i + fdow)%7].di_l3,
+                (int)days[(i + fdow)%7].di_l3,
                 days[(i + fdow)%7].di_name);
 	printf("\n");
     printf("===========================\n");
